get_lap_decomp_stats: cast eigenvalues to float32 (match eigvecs / model)
torch.from_numpy on float64 evals left EigVals as double; LapPE encoder
concatenates with float EigVecs and hit Linear Double vs Float in Exp2.

Made-with: Cursor

+++ graphgps/transform/posenc_stats.py
@@ -191,7 +191,8 @@ def get_lap_decomp_stats(evals, evects, max_freqs, eigvec_norm='L2'):
     # Keep up to the maximum desired number of frequencies.
     idx = evals.argsort()[:max_freqs]
     evals, evects = evals[idx], np.real(evects[:, idx])
-    evals = torch.from_numpy(np.real(evals)).clamp_min(0)
+    # float32 to match evects and model weights (from_numpy defaults to float64)
+    evals = torch.from_numpy(np.real(evals)).float().clamp_min(0)
 
     # Normalize and pad eigen vectors.
     evects = torch.from_numpy(evects).float()
